backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.12
